LayoutTests/http/tests/security/contentSecurityPolicy/worker-script-src-expected.txt

CONSOLE MESSAGE: Refused to load script from 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/alert-fail.js' because of Content-Security-Policy.

ALERT: Error: SECURITY_ERR: DOM Exception 18
